refactor(runtime): factor out `renderRichTextV2` into its own module
Factor out `renderRichTextV2` into its own module, mark the
corresponding client components with 'use client'
