+++ packages/runtime/src/runtimes/react/controls/rich-text-v2/render-rich-text-v2.tsx
@@ -1,0 +1,21 @@
+import { type ReactNode, lazy } from 'react'
+
+import { type DataType } from '@makeswift/controls'
+
+import { RichTextV2Control, RichTextV2Definition } from '../../../../controls/rich-text-v2'
+
+import { RichTextV2Value } from './rich-text-v2-value'
+
+const ReadOnlyTextV1 = lazy(() => import('../rich-text/ReadOnlyText'))
+
+export function renderRichTextV2(
+  data: DataType<RichTextV2Definition> | undefined,
+  definition: RichTextV2Definition,
+  control: RichTextV2Control | null,
+): ReactNode {
+  return RichTextV2Definition.isV1Data(data) ? (
+    <ReadOnlyTextV1 text={data} />
+  ) : (
+    <RichTextV2Value data={data} definition={definition} control={control} />
+  )
+}
