+++ packages/runtime/src/runtimes/react/controls/rich-text/ReadOnlyText.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { cx } from '@emotion/css'
 import { ForwardedRef, forwardRef } from 'react'
 import { Descendant, Text } from 'slate'
